% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova_lm}
\alias{anova_lm}
\title{build lm model between drugs accounting for batch effect}
\usage{
anova_lm(data, formula = "ratio ~ drug + batch", lower = TRUE, outdir,
  chunk = nrow(data))
}
\arguments{
\item{data}{a data.frame}

\item{formula}{(default: "ratio ~ drug + batch") the formula of the model}

\item{lower}{(default: TRUE) tests if "y" (the ratio) is lower than in controls}

\item{outdir}{set the outdir directory to a path (default extracted from fcs)}

\item{chunk}{(default: 20000) size of data chunk to do the computation on}
}
\value{
a data.frame
}
\description{
build lm model between drugs accounting for batch effect
}
\examples{
\dontrun{
data <- anova_lm(data)
}
}
