% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova_rlm}
\alias{anova_rlm}
\title{build rm model between drugs accounting for batch effect}
\usage{
anova_rlm(data, formula = "ratio ~ drug + batch")
}
\arguments{
\item{data}{a data.frame}

\item{formula}{(default: "ratio ~ drug + batch") the formula of the model}
}
\value{
a data.frame
}
\description{
build rm model between drugs accounting for batch effect
}
\examples{
\dontrun{
data <- anova_rlm(data)
}
}
