% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{CribleJurkat}
\alias{CribleJurkat}
\title{perform all the analysis for a CribleJurkat analysis}
\usage{
CribleJurkat(data_path, results_path)
}
\arguments{
\item{data_path}{path to the data folder}

\item{results_path}{path to the results folder}
}
\value{
TRUE if everythings ran correclty
}
\description{
perform all the analysis for a CribleJurkat analysis
}
\examples{
\dontrun{
CribleJurkat("data/2018_01_08_20171212/", 'results/2018_01_08_20171212/')
}
}
