% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{flowset2dataframe}
\alias{flowset2dataframe}
\title{convert annotated flowSet to data.frame}
\usage{
flowset2dataframe(fcs_data, channels = c("Y1.A", "B1.A"), norm = TRUE)
}
\arguments{
\item{fcs_data}{an object of class flowSet}

\item{channels}{(default = c("Y1.A", "B1.A)) the column ratio will be the
ratio of Y1.A / B1.A}

\item{norm}{(default: TRUE) normalization step (power_trans) for the ratio}
}
\value{
a data.frame
}
\description{
convert annotated flowSet to data.frame
}
\examples{
\dontrun{
data <- flowset2dataframe(fsc_data)
}
}
