% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_files}
\alias{get_files}
\title{get list of file from a folder}
\usage{
get_files(path, regexp)
}
\arguments{
\item{data_path}{path to the data folder}
}
\value{
list of file names
}
\description{
get list of file from a folder
}
\examples{
\dontrun{
get_files("data/examples/")
}
}
