% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_annotation}
\alias{load_annotation}
\title{read annotation.csv files in a folder}
\usage{
load_annotation(data_path)
}
\arguments{
\item{data_path}{path to the csv folder}
}
\value{
a data.frame
}
\description{
read annotation.csv files in a folder
}
\examples{
\dontrun{
load_annotation("data/examples/")
}
}
