% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{read fcs files in a folder}
\usage{
load_data(data_path)
}
\arguments{
\item{data_path}{path to the fcs folder}
}
\value{
an object of class flowSet
}
\description{
read fcs files in a folder
}
\examples{
\dontrun{
load_data("data/examples/")
}
}
