% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{mk_outdir}
\alias{mk_outdir}
\title{create outdir}
\usage{
mk_outdir(fcs_data, folder)
}
\arguments{
\item{fcs_dat}{an object of class flowSet}

\item{outdir}{a directory name}
}
\value{
the name of the project
}
\description{
create outdir
}
\examples{
\dontrun{
mk_outdir(fcs_data, "gating")
}
}
