% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{move_none_well}
\alias{move_none_well}
\title{convert data from anova_rlm to well position in chemical data base}
\usage{
move_none_well(data, col = c("01", "12"))
}
\arguments{
\item{data}{data.frame outputed by anova_rlm()}

\item{col}{(default: c("01", "12")) new position of the "None" columns}
}
\value{
data.frame with the None well in col
}
\description{
convert data from anova_rlm to well position in chemical data base
}
\examples{
\dontrun{
data_chem <- move_none_well(data)
}
}
