% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_column_ratio}
\alias{plot_column_ratio}
\title{plot the ratio distribution per column}
\usage{
plot_column_ratio(data, sample_size = nrow(data)/100, sufix = "")
}
\arguments{
\item{data}{data.frame object}

\item{sample_size}{(default: nrow(data) / 100) size of the subsample of data
to draw, to speedup the ploting proccess}

\item{file}{pdf file the save the object}
}
\value{
a ggplot2 object
}
\description{
plot the ratio distribution per column
}
\examples{
\dontrun{
plot_column_ratio(fsc_data)
}
}
