% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_well_ratio}
\alias{plot_well_ratio}
\title{plot the ratio distribution per well}
\usage{
plot_well_ratio(data, sample_size = nrow(data)/100, sufix = "")
}
\arguments{
\item{data}{data.frame object}

\item{sample_size}{(default: nrow(data) / 100) size of the subsample of data
to draw, to speedup the ploting proccess}

\item{sufix}{(default: "") sufix to append to pdf name}
}
\value{
a ggplot2 object
}
\description{
plot the ratio distribution per well
}
\examples{
\dontrun{
plot_plates_ratio(fsc_data)
}
}
