% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{project_name}
\alias{project_name}
\title{get project name}
\usage{
project_name(fcs_data)
}
\arguments{
\item{fcs_data}{an object of class flowSet}
}
\value{
the name of the project
}
\description{
get project name
}
\examples{
\dontrun{
project_name(fcs_data)
}
}
