% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{rm_debris}
\alias{rm_debris}
\title{remove debris from data}
\usage{
rm_debris(fcs_data)
}
\arguments{
\item{fcs_data}{an object of class flowSet}
}
\value{
an object of class flowSet
}
\description{
remove debris from data
}
\examples{
\dontrun{
fsc_data <- rm_debris(fsc_data)
}
}
