% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{rm_nonfluo}
\alias{rm_nonfluo}
\title{remove outliers on fluorecence channels}
\usage{
rm_nonfluo(fcs_data)
}
\arguments{
\item{fcs_data}{an object of class flowSet}
}
\value{
an object of class flowSet
}
\description{
remove outliers on fluorecence channels
}
\examples{
\dontrun{
fsc_data <- rm_nonfluo(fsc_data)
}
}
