% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{rm_nonsinglets}
\alias{rm_nonsinglets}
\title{remove non singlets}
\usage{
rm_nonsinglets(fcs_data)
}
\arguments{
\item{fcs_data}{an object of class flowSet}
}
\value{
an object of class flowSet
}
\description{
remove non singlets
}
\examples{
\dontrun{
fsc_data <- rm_nonsinglets(fsc_data)
}
}
