% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scaled_pval}
\alias{scaled_pval}
\title{plot the ratio distribution per well}
\usage{
scaled_pval(data)
}
\arguments{
\item{data}{data.frame object with a pval column}
}
\value{
data data.frame object with a scaled_pval column
}
\description{
plot the ratio distribution per well
}
\examples{
\dontrun{
data <- scaled_pval(data)
}
}
