% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{set_analysis}
\alias{set_analysis}
\title{perform all the analysis for a given data set}
\usage{
set_analysis(data_path = "data/", meta = F, output_data = F)
}
\arguments{
\item{data_path}{path to the data set}

\item{meta}{(default: FALSE) is the set_analysis part of a sets analysis}
}
\value{
TRUE if everythings ran correclty
}
\description{
perform all the analysis for a given data set
}
\examples{
\dontrun{
set_analysis("data/2018_01_08_20171212/")
}
}
