% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{split_lm}
\alias{split_lm}
\title{split lm model by chunck with biglm}
\usage{
split_lm(data, formula = "ratio ~ drug + batch", chunk = nrow(data))
}
\arguments{
\item{data}{a data.frame}

\item{formula}{(default: "ratio ~ drug + batch") the formula of the model}

\item{chunk}{(default: 200000) chunk size}
}
\value{
a data.frame
}
\description{
split lm model by chunck with biglm
}
