% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{analysis}
\alias{analysis}
\title{perform all the analysis for a set of data sets}
\usage{
analysis(data_path = "data/")
}
\arguments{
\item{data_path}{path to folder containing the data sets}
}
\value{
TRUE if everythings ran correclty
}
\description{
perform all the analysis for a set of data sets
}
\examples{
\dontrun{
analysis("data/set_test")
}
}
